User end view products 1

+++ sky-server/src/main/java/com/sky/mapper/DishFlavorMapper.java
@@ -22,7 +22,7 @@ public interface DishFlavorMapper {
      * @return dishVO
      */
     @Select("select * from dish_flavor where dish_id=#{id}")
-    List<DishFlavor> getFlavorList(Integer id);
+    List<DishFlavor> getFlavorList(Long id);
 
 
     void deleteBatch(List<Long> ids);
